UZEM: A line-buffer based renderer.
The line rendering is changed to write pixels into a 8 bit line buffer, which
is then shrunk down to the appropriate aspect ratio (4.66 cycles per pixel for
square aspect ratio).

The software shrinking is not too costly (it is about 2 percents of the CPU
cost of the emulator as of now), but it may also be a base for further
experiments moving this process onto the video card.

A notable improvement of this version versus the previous solution is the good
representation of higher resolution modes, even 3 cycles / pixel modes (code
tiles) look reasonably clean, not missing lines.

Note that it relies on SDL using some 8 bit per channel representation for the
32 bits / pixel software surface.