UZEM: First attempt at creating a renderer.
The primary goal was to create a proper interface through which multiple
renderer implementations may be supported, also removing some of the mess
within avr8.cpp.

Currently notably the capture feature is broken. It will likely be supported
by an additional interface function in the renderer to request lines of the
output.

Note that the capture module might want to create video at an arbitrary
frame rate not necessarily equaling that of the emulator, so a direct
connection is not preferred. A full surface pass is neither nice since the
renderer might have to convert, and the receiving capture module might be
streaming, so for this a single line of cached data is more optimal.

The full screen feature at least on my PC is also broken, but it rather seems
to be something with SDL than the code. SDL can not even restore the original
resolution when exiting.

Note: I enabled warnings in the Makefile! Please, please leave it there, only
disabling it for compiles! It helps cleaning up the mess!