UZEM: Refinements of the renderer interface.
Notably the renderer interface received a getLine method which can be used to
support capturing and creating screenshots. The latter is implemented and is
functional.

Minor improvements include the definition of a debug mode when the entire
NTSC signal output is visible. This is not implemented yet in the software
renderer (it will also need a proper implementation in avr8.cpp so all
parameters of the renderer's line method are properly set). A getProp method
was also added so if necessary the renderer's properties can be queried.

The software renderer's line generation was moved into a set of support
functions so other renderers might also use it.

Files avr8.cpp and Makefile have consistent Microsoft line endings, new files
related to the renderer have Unix line endings.